+++ CHANGELOG.md
@@ -1,6 +1,10 @@
 ## 0.2.0
 #### Added 
 - Edit Squad OKRs
+- Added proper squad and OKR count visualizations to UserList
+
+#### Removed
+- Hid User Details link from users page
 
 #### Fixed
 - When there are no checkins, output is `NaN` instead of `0`